+++ riscv-peripheral/examples/e310x.rs
@@ -2,35 +2,18 @@
 //! This is a simple example of how to use the `riscv-peripheral` crate to generate
 //! peripheral definitions for a target.
 
-use riscv_pac::{HartIdNumber, InterruptNumber, PriorityNumber};
+use riscv_pac::pac_enum;
 
 #[repr(u16)]
+#[pac_enum(unsafe HartIdNumber)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum HartId {
     H0 = 0,
 }
 
-unsafe impl HartIdNumber for HartId {
-    const MAX_HART_ID_NUMBER: u16 = 0;
-
-    #[inline]
-    fn number(self) -> u16 {
-        self as _
-    }
-
-    #[inline]
-    fn from_number(number: u16) -> Result<Self, u16> {
-        if number > Self::MAX_HART_ID_NUMBER {
-            Err(number)
-        } else {
-            // SAFETY: valid context number
-            Ok(unsafe { core::mem::transmute(number) })
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 #[repr(u16)]
+#[pac_enum(unsafe ExternalInterruptNumber)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum Interrupt {
     WATCHDOG = 1,
     RTC = 2,
@@ -86,27 +69,9 @@ pub enum Interrupt {
     I2C0 = 52,
 }
 
-unsafe impl InterruptNumber for Interrupt {
-    const MAX_INTERRUPT_NUMBER: u16 = 52;
-
-    #[inline]
-    fn number(self) -> u16 {
-        self as _
-    }
-
-    #[inline]
-    fn from_number(number: u16) -> Result<Self, u16> {
-        if number == 0 || number > Self::MAX_INTERRUPT_NUMBER {
-            Err(number)
-        } else {
-            // SAFETY: valid interrupt number
-            Ok(unsafe { core::mem::transmute(number) })
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 #[repr(u8)]
+#[pac_enum(unsafe PriorityNumber)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum Priority {
     P0 = 0,
     P1 = 1,
@@ -116,25 +81,6 @@ pub enum Priority {
     P5 = 5,
     P6 = 6,
     P7 = 7,
-}
-
-unsafe impl PriorityNumber for Priority {
-    const MAX_PRIORITY_NUMBER: u8 = 7;
-
-    #[inline]
-    fn number(self) -> u8 {
-        self as _
-    }
-
-    #[inline]
-    fn from_number(number: u8) -> Result<Self, u8> {
-        if number > Self::MAX_PRIORITY_NUMBER {
-            Err(number)
-        } else {
-            // SAFETY: valid priority number
-            Ok(unsafe { core::mem::transmute(number) })
-        }
-    }
 }
 
 #[cfg(feature = "aclint-hal-async")]