+++ riscv-peripheral/src/plic.rs
@@ -144,10 +144,11 @@ impl<P: Plic> CTX<P> {
 
 #[cfg(test)]
 pub(crate) mod test {
-    use super::{HartIdNumber, InterruptNumber, PriorityNumber};
+    use riscv_pac::*;
 
+    #[pac_enum(unsafe ExternalInterruptNumber)]
     #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-    #[repr(u16)]
+    #[repr(usize)]
     pub(crate) enum Interrupt {
         I1 = 1,
         I2 = 2,
@@ -155,6 +156,7 @@ pub(crate) mod test {
         I4 = 4,
     }
 
+    #[pac_enum(unsafe PriorityNumber)]
     #[derive(Clone, Copy, Debug, Eq, PartialEq)]
     #[repr(u8)]
     pub(crate) enum Priority {
@@ -164,69 +166,13 @@ pub(crate) mod test {
         P3 = 3,
     }
 
+    #[pac_enum(unsafe HartIdNumber)]
     #[derive(Clone, Copy, Debug, Eq, PartialEq)]
     #[repr(u16)]
     pub(crate) enum Context {
         C0 = 0,
         C1 = 1,
         C2 = 2,
-    }
-
-    unsafe impl InterruptNumber for Interrupt {
-        const MAX_INTERRUPT_NUMBER: u16 = 4;
-
-        #[inline]
-        fn number(self) -> u16 {
-            self as _
-        }
-
-        #[inline]
-        fn from_number(number: u16) -> Result<Self, u16> {
-            if number > Self::MAX_INTERRUPT_NUMBER || number == 0 {
-                Err(number)
-            } else {
-                // SAFETY: valid interrupt number
-                Ok(unsafe { core::mem::transmute(number) })
-            }
-        }
-    }
-
-    unsafe impl PriorityNumber for Priority {
-        const MAX_PRIORITY_NUMBER: u8 = 3;
-
-        #[inline]
-        fn number(self) -> u8 {
-            self as _
-        }
-
-        #[inline]
-        fn from_number(number: u8) -> Result<Self, u8> {
-            if number > Self::MAX_PRIORITY_NUMBER {
-                Err(number)
-            } else {
-                // SAFETY: valid priority number
-                Ok(unsafe { core::mem::transmute(number) })
-            }
-        }
-    }
-
-    unsafe impl HartIdNumber for Context {
-        const MAX_HART_ID_NUMBER: u16 = 2;
-
-        #[inline]
-        fn number(self) -> u16 {
-            self as _
-        }
-
-        #[inline]
-        fn from_number(number: u16) -> Result<Self, u16> {
-            if number > Self::MAX_HART_ID_NUMBER {
-                Err(number)
-            } else {
-                // SAFETY: valid context number
-                Ok(unsafe { core::mem::transmute(number) })
-            }
-        }
     }
 
     #[test]